Code smell: Long function that can be broken into 2.
I think I should point out that this is nitpicky, as reading the log is only used by one function, and therefore breaking that into its own function is a bit pointless. Im more making this change to furfull assingment requirements, though this would be a good change if functionality was going to continue being built as the function could benifit a future endeavor.